lib/shadow/subid/: sgetsient(): Add function
